Code to insert and retrieve player details from database

+++ app/src/main/java/com/boardgames/jaipur/entities/Player.java
@@ -1,7 +1,5 @@
 package com.boardgames.jaipur.entities;
 
-import android.graphics.Bitmap;
-
 import androidx.annotation.NonNull;
 import androidx.room.ColumnInfo;
 import androidx.room.Entity;
@@ -11,7 +9,7 @@ import androidx.room.PrimaryKey;
 public class Player {
 
     @PrimaryKey(autoGenerate = true)
-    private int id;
+    private long id;
 
     @NonNull
     @ColumnInfo(name = "name")
@@ -47,11 +45,11 @@ public class Player {
         this.timeUpdated = timeUpdated;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
